+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Version 1.1.0]
+### Added
+- New field status
+
+### Changed
+- Add select status
+- Results style
+
 ## [Version 1.0.0]
 ### Added
 - Admin CRUD